made changes

+++ CSD2c/Prototype/proto_eind_1/filter.h
@@ -11,5 +11,7 @@ public:
 
 protected:
     uint msToSamples(float delayMS);
-    float m_samplerate;
+    double m_samplerate;
+
+
 };
